// apps/admin-web/src/api/model/staffResponseDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Admin API
 * The Admin API description
 * OpenAPI spec version: 1.0
 */
import type { StaffResponseDtoRole } from './staffResponseDtoRole';
import type { StaffResponseDtoStatus } from './staffResponseDtoStatus';
import type { StaffTermAgreementResponseDto } from './staffTermAgreementResponseDto';
import type { StaffUserResponseDto } from './staffUserResponseDto';

export interface StaffResponseDto {
  id: string;
  role: StaffResponseDtoRole;
  user: StaffUserResponseDto;
  status: StaffResponseDtoStatus;
  createdAt: string;
  updatedAt: string;
  termAgreements?: StaffTermAgreementResponseDto[];
}